requirements.txt
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
librosa>=0.10.2
scipy>=1.14.1
matplotlib>=3.9.3
numpy>=2.1.0
google-generativeai==0.8.3
python-dotenv==1.0.1
Pillow>=11.1.0
torch
tensorflow
tensorflow-hub
pandas
